+++ projects/angular2-smart-table/src/lib/components/thead/thead.module.ts
@@ -9,7 +9,6 @@ import {TheadCreateCancelComponent} from './cells/create-cancel.component';
 import {ActionsTitleComponent} from './cells/actions-title.component';
 import {AddButtonComponent} from './cells/add-button.component';
 import {ColumnTitleComponent} from './cells/column-title.component';
-import {TitleComponent} from './cells/title/title.component';
 import {TheadFitlersRowComponent} from './rows/thead-filters-row.component';
 import {TheadFormRowComponent} from './rows/thead-form-row.component';
 import {TheadTitlesRowComponent} from './rows/thead-titles-row.component';
@@ -21,7 +20,6 @@ const THEAD_COMPONENTS = [
   ActionsTitleComponent,
   AddButtonComponent,
   ColumnTitleComponent,
-  TitleComponent,
   TheadFitlersRowComponent,
   TheadFormRowComponent,
   TheadTitlesRowComponent,