+++ projects/angular2-smart-table/CHANGELOG.md
@@ -4,6 +4,8 @@ This document lists the changes introduced by this fork.
 
 ## Version 3.6.0
 
+* Improve rendering of the column title header
+
 ## Version 3.5.2
 
 * Fix that after removing the unnecessary paging events, the first page was not selected anymore when data was lazy loaded